Lighting bug Fixed, Lighting works of pixel values now and added secure paths

+++ RPGTeamGame/src/Main/Background/MapGen/Block.java
@@ -37,9 +37,11 @@ public class Block
         return collidable;
     }
 
-    public Vector2i getPosition() {
-        return new Vector2i(xPosition, yPosition);
+    public Vector2f getPosition() {
+        return new Vector2f(xPosition, yPosition);
     }
+
+
 
     public Vector2i inBlock()
     {
